test: deflake test-buffer-large-size
